refactor: move parseLatex utility function to LatexToSympy.ts
Will allow this function to be called from parser, needed for parametric plot capability
